+++ test/test.config.js
@@ -5,12 +5,16 @@
 
 const {config} = require('bedrock');
 const path = require('path');
+require('bedrock-mongodb');
+require('bedrock-express');
 
 config.mocha.tests.push(path.join(__dirname, 'mocha'));
 
 // MongoDB
 config.mongodb.name = 'bedrock_account_http_test';
-config.mongodb.local.collection = 'bedrock_account_http_test';
 config.mongodb.dropCollections = {};
 config.mongodb.dropCollections.onInit = true;
 config.mongodb.dropCollections.collections = [];
+
+// enable sessions
+config.express.useSession = true;